Switch server from mod_wsgi to gunicorn
The primary change in this commit is the transition from using mod_wsgi as the server to gunicorn. The server command has been adjusted in the startup shell script to start gunicorn instead of mod_wsgi. To support this, gunicorn has been added to the project requirements. Additionally, a new status choice "pending-approval" has been added to the system status state choices in the common models.